Restructure neural module
