+++ src/ott/neural/models/layers.py
@@ -13,9 +13,9 @@
 # limitations under the License.
 from typing import Any, Callable, Tuple
 
-import flax.linen as nn
 import jax
 import jax.numpy as jnp
+from flax import linen as nn
 
 __all__ = ["PositiveDense", "PosDefPotentials"]
