+++ src/ott/solvers/__init__.py
@@ -11,10 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import contextlib
-
 from . import linear, quadratic, was_solver
-
-with contextlib.suppress(ImportError):
-  from . import nn
-del contextlib
