+++ src/ott/tools/__init__.py
@@ -11,8 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import contextlib
-
 from . import (
     gaussian_mixture,
     k_means,
@@ -21,7 +19,3 @@ from . import (
     sinkhorn_divergence,
     soft_sort,
 )
-
-with contextlib.suppress(ImportError):
-  from . import map_estimator
-del contextlib
