+++ tests/neural/__init__.py
@@ -1,0 +1,3 @@
+import pytest
+
+_ = pytest.importorskip("flax")
