+++ tests/neural/icnn_test.py
@@ -15,10 +15,7 @@ import jax
 import jax.numpy as jnp
 import numpy as np
 import pytest
-
-_ = pytest.importorskip("flax")
-
-from ott.solvers.nn import models
+from ott.neural.models import models
 
 
 @pytest.mark.fast()
