+++ tests/neural/losses_test.py
@@ -15,11 +15,9 @@
 import jax
 import numpy as np
 import pytest
-
-_ = pytest.importorskip("flax")
-
 from ott.geometry import costs
-from ott.solvers.nn import losses, models
+from ott.neural.models import models
+from ott.neural.solvers import losses
 
 
 @pytest.mark.fast()
